Move buf image generation
