--- pkg/star/buf.go
@@ -1,52 +1,0 @@
-// Copyright 2022 Lekko Technologies, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package star
-
-import (
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-
-	"github.com/pkg/errors"
-)
-
-type bufImage struct {
-	filename string
-}
-
-func newBufImage(protoDir string) (*bufImage, error) {
-	outputFile := filepath.Join(protoDir, "image.bin")
-	args := []string{
-		"build",
-		protoDir,
-		"--exclude-source-info",
-		fmt.Sprintf("-o %s", outputFile),
-	}
-	cmd := exec.Command("buf", args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err := cmd.Run()
-	if err != nil {
-		return nil, errors.Wrap(err, "buf build")
-	}
-	return &bufImage{
-		filename: outputFile,
-	}, nil
-}
-
-func (bi *bufImage) cleanup() error {
-	return os.Remove(bi.filename)
-}
